(Line colors) Add colors to lines and palette support
This makes the lines less confusing when nodeCount is high.

A gradient array will be generated on startup, based on some colors of a palette.
The number of colors in the gradient array will be equal to the maximum number of nodes
there can be, but hen, also on startup or when nodeCount changes, some colors will be chosen
to be used for the lines, more or less like a "less resolution gradient".