Fix scrollable palette item bug
Firefox didn't load the widths properly at first so all items with 5 colors or more
were given the .scrollable class. Then, because the window hadn't been resized, it didn't update.
By using a useEffect hook that updates when clientWidth changes, all items work properly, as when the width
is loaded, the .scrollable class is updated.